EDATOS-3294 - Consulta de favoritos
> Mejora el rendimiento al obtener los usuarios suscritos a cada tema.

+++ edatos-external-users-core/src/main/java/es/gobcan/istac/edatos/external/users/core/service/FavoriteService.java
@@ -19,5 +19,5 @@ public interface FavoriteService {
     Page<FavoriteEntity> find(String query, Pageable pageable);
     List<FavoriteEntity> find(String query, Sort sort);
     void delete(FavoriteEntity favorite);
-    int getCategorySubscribers(CategoryEntity category);
+    long getCategorySubscribers(CategoryEntity category);
 }
