+++ edatos-external-users-rest-common/src/main/java/es/gobcan/istac/edatos/external/users/rest/common/dto/CategoryDto.java
@@ -23,7 +23,7 @@ public class CategoryDto extends AuditableDto implements Identifiable {
     private Instant updateDate;
     private Instant lastUpdated;
     private String uuid;
-    private int subscribedUsers;
+    private long subscribedUsers;
     private List<CategoryDto> children = new ArrayList<>();
 
     @Override
@@ -135,11 +135,11 @@ public class CategoryDto extends AuditableDto implements Identifiable {
         this.children = children;
     }
 
-    public int getSubscribedUsers() {
+    public long getSubscribedUsers() {
         return subscribedUsers;
     }
 
-    public void setSubscribedUsers(int subscribedUsers) {
+    public void setSubscribedUsers(long subscribedUsers) {
         this.subscribedUsers = subscribedUsers;
     }
 }
